Refactor to adapt paas-app-charmer for new frameworks
