+++ .trivyignore
@@ -4,3 +4,5 @@ CVE-2024-24790
 CVE-2023-45288
 # ignore CVE introduced by python3-gunicorn
 CVE-2022-40897
+# pypa/setuptools: Remote code execution via download
+CVE-2024-6345
