+++ paas_app_charmer/django/__init__.py
@@ -1,9 +1,8 @@
-#!/usr/bin/env python3
-
 # Copyright 2024 Canonical Ltd.
 # See LICENSE file for licensing details.
 
 """Django module entrypoint."""
 
-# flake8 complains imported but unused
-from .charm import Charm  # noqa: disable=F401
+from .charm import Charm
+
+__all__ = ["Charm"]
