+++ paas_app_charmer/django/charm.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python3
-
 # Copyright 2024 Canonical Ltd.
 # See LICENSE file for licensing details.
 
@@ -10,19 +8,14 @@ import secrets
 import typing
 
 import ops
-
-# pydantic is causing this no-name-in-module problem
-from pydantic import BaseModel, Extra, Field, ValidationError  # pylint: disable=no-name-in-module
+from pydantic import BaseModel, Extra, Field, validator
 
 from paas_app_charmer._gunicorn.charm import GunicornBase
-from paas_app_charmer._gunicorn.charm_utils import block_if_invalid_config
-from paas_app_charmer.exceptions import CharmConfigInvalidError
-from paas_app_charmer.utils import build_validation_error_message
 
 logger = logging.getLogger(__name__)
 
 
-class DjangoConfig(BaseModel, extra=Extra.allow):  # pylint: disable=too-few-public-methods
+class DjangoConfig(BaseModel, extra=Extra.ignore):
     """Represent Django builtin configuration values.
 
     Attrs:
@@ -32,13 +25,34 @@ class DjangoConfig(BaseModel, extra=Extra.allow):  # pylint: disable=too-few-pub
         allowed_hosts: a list of host/domain names that this Django site can serve.
     """
 
-    debug: bool | None = Field(default=None)
-    secret_key: str | None = Field(default=None, min_length=1)
-    allowed_hosts: list[str]
+    debug: bool | None = Field(alias="django-debug", default=None)
+    secret_key: str | None = Field(alias="django-secret-key", default=None, min_length=1)
+    allowed_hosts: str | None = Field(alias="django-allowed-hosts", default=[])
+
+    @validator("allowed_hosts")
+    @classmethod
+    def allowed_hosts_to_list(cls, value: str | None) -> typing.List[str]:
+        """Convert a comma separated list of allowed hosts to list.
+
+        Args:
+          value: allowed hosts as string.
+
+        Return:
+          list of allowed hosts.
+        """
+        if not value:
+            return []
+        return [h.strip() for h in value.split(",")]
 
 
-class Charm(GunicornBase):  # pylint: disable=too-many-instance-attributes
-    """Django Charm service."""
+class Charm(GunicornBase):
+    """Django Charm service.
+
+    Attrs:
+        framework_config_class: Base class for framework configuration.
+    """
+
+    framework_config_class = DjangoConfig
 
     def __init__(self, framework: ops.Framework) -> None:
         """Initialize the Django charm.
@@ -46,35 +60,8 @@ class Charm(GunicornBase):  # pylint: disable=too-many-instance-attributes
         Args:
             framework: operator framework.
         """
-        super().__init__(framework=framework, wsgi_framework="django")
-        self.framework.observe(self.on.django_app_pebble_ready, self._on_django_app_pebble_ready)
+        super().__init__(framework=framework, framework_name="django")
         self.framework.observe(self.on.create_superuser_action, self._on_create_superuser_action)
-
-    def get_wsgi_config(self) -> BaseModel:
-        """Return Django framework related configurations.
-
-        Returns:
-             Django framework related configurations.
-
-        Raises:
-            CharmConfigInvalidError: if charm config is not valid.
-        """
-        django_config: dict[str, typing.Any] = {
-            "debug": self.config.get("django-debug"),
-            "secret_key": self.config.get("django-secret-key"),
-        }
-        allowed_hosts = str(self.config.get("django-allowed-hosts", ""))
-        if allowed_hosts.strip():
-            django_config["allowed_hosts"] = [h.strip() for h in allowed_hosts.split(",")]
-        else:
-            django_config["allowed_hosts"] = []
-        try:
-            return DjangoConfig.model_validate(django_config)
-        except ValidationError as exc:
-            error_message = build_validation_error_message(
-                exc, prefix="django-", underscore_to_dash=True
-            )
-            raise CharmConfigInvalidError(f"invalid configuration: {error_message}") from exc
 
     def get_cos_dir(self) -> str:
         """Return the directory with COS related files.
@@ -83,11 +70,6 @@ class Charm(GunicornBase):  # pylint: disable=too-many-instance-attributes
             Return the directory with COS related files.
         """
         return str((pathlib.Path(__file__).parent / "cos").absolute())
-
-    @block_if_invalid_config
-    def _on_django_app_pebble_ready(self, _: ops.PebbleReadyEvent) -> None:
-        """Handle the pebble-ready event."""
-        self.restart()
 
     def _on_create_superuser_action(self, event: ops.ActionEvent) -> None:
         """Handle the create-superuser action.
