+++ tests/conftest.py
@@ -6,7 +6,7 @@
 
 def pytest_addoption(parser):
     """Define some command line options for integration and unit tests."""
-    parser.addoption("--charm-file", action="store")
+    parser.addoption("--charm-file", action="extend", nargs="+", default=[])
     parser.addoption("--test-flask-image", action="store")
     parser.addoption("--test-db-flask-image", action="store")
     parser.addoption("--django-app-image", action="store")
