+++ tests/integration/django/test_django_integrations.py
