+++ tests/unit/flask/test_database_migration.py
@@ -8,10 +8,10 @@ import ops
 import pytest
 from ops.testing import Harness
 
-from paas_app_charmer._gunicorn.charm_state import CharmState
 from paas_app_charmer._gunicorn.webserver import GunicornWebserver, WebserverConfig
-from paas_app_charmer._gunicorn.workload_config import WorkloadConfig
+from paas_app_charmer._gunicorn.workload_config import create_workload_config
 from paas_app_charmer._gunicorn.wsgi_app import WsgiApp
+from paas_app_charmer.charm_state import CharmState
 from paas_app_charmer.database_migration import DatabaseMigration, DatabaseMigrationStatus
 from paas_app_charmer.exceptions import CharmConfigInvalidError
 
@@ -35,9 +35,7 @@ def test_database_migration(harness: Harness):
         is_secret_storage_ready=True,
         secret_key="",
     )
-    workload_config = WorkloadConfig(
-        framework="flask",
-    )
+    workload_config = create_workload_config(framework_name="flask")
     webserver_config = WebserverConfig()
     webserver = GunicornWebserver(
         webserver_config=webserver_config,
@@ -111,9 +109,7 @@ def test_database_migrate_command(harness: Harness, file: str, command: list[str
         secret_key="",
     )
     webserver_config = WebserverConfig()
-    workload_config = WorkloadConfig(
-        framework="flask",
-    )
+    workload_config = create_workload_config(framework_name="flask")
     webserver = GunicornWebserver(
         webserver_config=webserver_config,
         workload_config=workload_config,
